+++ passkey-types/src/webauthn/extensions/mod.rs
@@ -2,8 +2,10 @@ use serde::{Deserialize, Serialize};
 use typeshare::typeshare;
 
 mod credential_properties;
+mod pseudo_random_function;
 
 pub use credential_properties::*;
+pub use pseudo_random_function::*;
 
 /// This is a dictionary containing the client extension input values for zero or more
 /// [WebAuthn Extensions]. There are currently none supported.
@@ -11,7 +13,7 @@ pub use credential_properties::*;
 /// <https://w3c.github.io/webauthn/#dictdef-authenticationextensionsclientinputs>
 ///
 /// [WebAuthn Extensions]: https://w3c.github.io/webauthn/#webauthn-extensions
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Default, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
 #[typeshare]
 pub struct AuthenticationExtensionsClientInputs {
@@ -20,6 +22,12 @@ pub struct AuthenticationExtensionsClientInputs {
     /// See [`CredentialPropertiesOutput`] for more information.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub cred_props: Option<bool>,
+
+    /// Inputs for the pseudo-random function extensions.
+    ///
+    /// See [`AuthenticationExtensionsPrfInputs`] for more information.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub prf: Option<AuthenticationExtensionsPrfInputs>,
 }
 
 /// This is a dictionary containing the client extension output values for zero or more
@@ -31,10 +39,16 @@ pub struct AuthenticationExtensionsClientInputs {
 #[derive(Debug, Default, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
 #[typeshare]
-pub struct AuthenticatorExtensionsClientOutputs {
+pub struct AuthenticationExtensionsClientOutputs {
     /// Contains properties of the given [`PublicKeyCredential`] when it is included.
     ///
     /// See [`CredentialPropertiesOutput`] for more information
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub cred_props: Option<CredentialPropertiesOutput>,
+
+    /// Contains the results of evaluating the PRF.
+    ///
+    /// See [`AuthenticationExtensionsPrfOutputs`] for more information.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub prf: Option<AuthenticationExtensionsPrfOutputs>,
 }